feat(models): change object_id and subscription_id fields to BigIntegerField

+++ strava/models.py
@@ -274,10 +274,10 @@ class Event(models.Model):
 
     aspect_type = models.CharField(max_length=128, choices=ASPECT_TYPES)
     event_time = models.DateTimeField()
-    object_id = models.IntegerField()
+    object_id = models.BigIntegerField()
     object_type = models.CharField(max_length=128, choices=OBJECT_TYPES)
     owner_id = models.ForeignKey(Runner, on_delete=models.CASCADE, related_name="updates", to_field="strava_id")
-    subscription_id = models.IntegerField()
+    subscription_id = models.BigIntegerField()
     updates = models.JSONField(default=dict)
     processed = models.BooleanField(default=False)
